Split online_serving.py for cleaner dependencies
